Node related code moved into `node` module
